fix: remove short arg, add arguments to the output
