+++ apps/vsce/src/components/engineService.ts
@@ -304,7 +304,7 @@ export class EngineService {
 	public async __getCodemodNames(): Promise<ReadonlyArray<string>> {
 		const childProcess = spawn(
 			CODEMOD_ENGINE_NODE_COMMAND,
-			["list", "--json", "--short"],
+			["list", "--json"],
 			{
 				stdio: "pipe",
 				shell: true,
