+++ packages/utilities/src/index.ts
@@ -16,6 +16,7 @@ export {
 	type CodemodConfigInput,
 	type KnownEngines,
 } from "./schemata/codemodConfigSchema.js";
+export { type CodemodListReturn } from "./schemata/codemodListResponse.js";
 export {
 	JOB_KIND,
 	parseSurfaceAgnosticJob,
