+++ packages/utilities/src/schemata/codemodListResponse.ts
@@ -1,0 +1,10 @@
+import { AllEngines, Arguments } from "./codemodConfigSchema.js";
+
+export type CodemodListReturn = {
+	name: string;
+	author: string;
+	engine: AllEngines;
+	tags: string[];
+	verified: boolean;
+	arguments: Arguments;
+}[];
